src/main/java/com/joelmaciel/agiledevprojects/domain/services/impl/CompanyServiceImpl.java
package com.joelmaciel.agiledevprojects.domain.services.impl;

import com.joelmaciel.agiledevprojects.api.dtos.CompanyDTO;
import com.joelmaciel.agiledevprojects.api.dtos.CompanyRequest;
import com.joelmaciel.agiledevprojects.domain.entities.Company;
import com.joelmaciel.agiledevprojects.domain.repositories.CompanyRepository;
import com.joelmaciel.agiledevprojects.domain.services.CompanyService;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@RequiredArgsConstructor
public class CompanyServiceImpl implements CompanyService {

    private final CompanyRepository companyRepository;

    @Override
    @Transactional
    public CompanyDTO save(CompanyRequest companyRequest) {
        Company company = CompanyRequest.toEntity(companyRequest);
        return CompanyDTO.toDTO(companyRepository.save(company));
    }
}
